+++ CHANGELOG.md
@@ -1,4 +1,6 @@
+## 1.1.0
+ * Added ability to use hours, minutes and seconds in a 'duration'
 ## 1.0.2
  * Fixed tooltip date format not respected
 ## 1.0.1
-* Fixed start date calculation
+ * Fixed start date calculation